requirements.txt
Flask
Flask-Cors
requests
PyYAML
CacheControl
gunicorn
Pyrebase4
google-generativeai
firebase-admin
bcrypt
google-auth-oauthlib
Markdown
parsel
Google-Images-Search
Unidecode
#windows-curses
curses-util
youtube-search-python
cloudmersive-barcode-api-client
qreader
opencv-python
